# Documentazione/apidocs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /admin/speakers/{id}:
    put:
      tags:
      - Speakers
      summary: Modifica un relatore esistente
      description: Solo ADMIN.
      operationId: updateSpeaker
      parameters:
      - name: id
        in: path
        description: ID relatore
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakerRequestDTO"
        required: true
      responses:
        "200":
          description: Relatore aggiornato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpeakerResponseDTO"
        "400":
          description: Dati non validi
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpeakerResponseDTO"
        "404":
          description: Relatore non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpeakerResponseDTO"
    delete:
      tags:
      - Speakers
      summary: Elimina un relatore
      description: Solo ADMIN.
      operationId: deleteSpeaker
      parameters:
      - name: id
        in: path
        description: ID relatore
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Relatore eliminato
        "404":
          description: Relatore non trovato
...
  /events/{eventId}/book:
    post:
      tags:
      - Tickets
      summary: Prenota un biglietto
      description: |
        Prenota un biglietto (STANDARD o VIP) per l'evento specificato.
        Il prezzo viene impostato automaticamente in base al tipo scelto.
        Vincoli: evento futuro, nessuna doppia prenotazione, posti disponibili, utente non bannato.
      operationId: bookTicket
      parameters:
      - name: eventId
        in: path
        description: ID evento
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketRequestDTO"
        required: true
      responses:
        "201":
          description: Biglietto prenotato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketResponseDTO"
        "400":
          description: Tipo biglietto non valido
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketResponseDTO"
        "404":
          description: Evento non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketResponseDTO"
        "409":
          description: "Evento passato, doppia prenotazione, posti esauriti o utente\
            \ bannato"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketResponseDTO"
...
  /admin/tags:
    post:
      tags:
      - Tags
      summary: Crea un nuovo tag
      description: Solo ADMIN.
      operationId: createTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagRequestDTO"
        required: true
      responses:
        "201":
          description: Tag creato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TagResponseDTO"
        "400":
          description: Dati non validi
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TagResponseDTO"
        "409":
          description: Tag con lo stesso nome già esistente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TagResponseDTO"
  /admin/speakers:
    post:
      tags:
      - Speakers
      summary: Crea un nuovo relatore
      description: Solo ADMIN.
      operationId: createSpeaker
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakerRequestDTO"
        required: true
      responses:
        "201":
          description: Relatore creato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpeakerResponseDTO"
        "400":
          description: Dati non validi
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpeakerResponseDTO"
...
  /tickets/my:
    get:
      tags:
      - Tickets
      summary: Le mie prenotazioni
      description: Lista di tutti i biglietti dell'utente autenticato.
      operationId: getMyTickets
      responses:
        "200":
          description: Lista biglietti
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketResponseDTO"
        "401":
          description: Non autenticato
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketResponseDTO"
...
  /speakers/{id}:
    get:
      tags:
      - Speakers
      summary: Dettaglio relatore per ID
      operationId: getSpeakerById
      parameters:
      - name: id
        in: path
        description: ID relatore
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Relatore trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpeakerResponseDTO"
        "404":
          description: Relatore non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpeakerResponseDTO"
...
  /tickets/{id}:
    delete:
      tags:
      - Tickets
      summary: Cancella una prenotazione
      description: Cancellazione consentita solo prima dell'inizio dell'evento. Il
        biglietto viene marcato come CANCELLED.
      operationId: cancelTicket
      parameters:
      - name: id
        in: path
        description: ID biglietto
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Prenotazione cancellata
        "403":
          description: Il biglietto non appartiene all'utente
        "404":
          description: Biglietto non trovato
        "409":
          description: Biglietto già cancellato o evento già iniziato
  /admin/tags/{id}:
    delete:
      tags:
      - Tags
      summary: Elimina un tag
      description: Solo ADMIN.
      operationId: deleteTag
      parameters:
      - name: id
        in: path
        description: ID tag
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Tag eliminato
        "404":
          description: Tag non trovato
...
components:
  schemas:
...
    SpeakerResponseDTO:
      type: object
      description: Dati di un relatore
      properties:
        id:
          type: integer
          format: int32
          description: ID relatore
          example: 1
        firstName:
          type: string
          description: Nome
          example: Mario
        lastName:
          type: string
          description: Cognome
          example: Rossi
        bio:
          type: string
          description: Biografia
          example: Senior developer con 10 anni di esperienza
        photoUrl:
          type: string
          description: URL foto
          example: https://example.com/foto.jpg
    TagResponseDTO:
      type: object
      description: Dati di un tag
      properties:
        id:
          type: integer
          format: int32
          description: ID tag
          example: 1
        name:
          type: string
          description: Nome del tag
          example: Java
...
    SpeakerRequestDTO:
      type: object
      description: Dati per la creazione o modifica di un relatore
      properties:
        firstName:
          type: string
          description: Nome
          example: Mario
          minLength: 1
        lastName:
          type: string
          description: Cognome
          example: Rossi
          minLength: 1
        bio:
          type: string
          description: Biografia
          example: Senior developer con 10 anni di esperienza
        photoUrl:
          type: string
          description: URL foto
          example: https://example.com/foto.jpg
      required:
      - firstName
      - lastName
...
    TicketRequestDTO:
      type: object
      description: Dati per la prenotazione di un biglietto
      properties:
        type:
          type: string
          description: Tipo di biglietto
          enum:
          - STANDARD
          - VIP
          example: STANDARD
        eventId:
          type: integer
          format: int32
          description: ID dell'evento
          example: 1
      required:
      - eventId
      - type
    TicketResponseDTO:
      type: object
      description: Dati di un biglietto prenotato
      properties:
        id:
          type: integer
          format: int32
          description: ID biglietto
          example: 1
        type:
          type: string
          description: Tipo di biglietto
          enum:
          - STANDARD
          - VIP
          example: STANDARD
        price:
          type: number
          description: Prezzo pagato
          example: 49.99
        status:
          type: string
          description: Stato del biglietto
          enum:
          - ACTIVE
          - CANCELLED
          example: ACTIVE
        eventId:
          type: integer
          format: int32
          description: ID evento
          example: 1
        eventTitle:
          type: string
          description: Titolo evento
          example: Workshop Spring Boot
...
    TagRequestDTO:
      type: object
      description: Dati per la creazione di un tag
      properties:
        name:
          type: string
          description: Nome del tag
          example: Java
          minLength: 1
      required:
      - name